preliminary unix socket support
If the host passed to connect() starts with a "/", then a connection
attempt is made to the unix socket at the specified path.  The port
argument is currently still required, but will be ignored.  Attempting to
connect to a unix socket on Windows will raise an AttributeError (because
socket.AF_UNIX isn't defined), however this will later be changed.