mpd.py: check if iterating before fetching response
If an iterator is currently in progress, all functions which fetch a
response (such as <cmd>() and fetch_<cmd>()) will be disabled, as well as
command lists.  This makes iterators much more reliable by preventing code
from reading a response while an iterator is trying to read another
response.  Previously this was allowed, and would corrupt the library state
by providing false responses to both the iterator and the code trying to
fetch a different response.